requirements.txt
altgraph==0.17
astroid==2.4.2
autopep8==1.5.4
certifi==2020.11.8
cycler==0.10.0
greenlet==0.4.17
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
macholib==1.14
matplotlib==3.4.2
mccabe==0.6.1
modulegraph==0.18
msgpack==1.0.0
numpy==1.22.1
pandas==1.1.5
PeakUtils==1.3.3
Pillow==8.3.2
py2app==0.23
pycodestyle==2.6.0
pylint==2.6.0
pynvim==0.4.2
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyqtgraph==0.11.0
python-dateutil==2.8.1
pytz==2020.4
SciencePlots==1.0.9
scipy==1.5.4
six==1.15.0
tenacity==8.0.1
toml==0.10.2
wrapt==1.12.1
